refactor: :recycle: Extract account verification to another service to avoid circular dependency
